Move InputFields to account common module
