+++ feature/account/common/src/main/kotlin/app/k9mail/feature/account/common/domain/input/BooleanInputField.kt
@@ -1,49 +1,48 @@
-package app.k9mail.feature.account.setup.domain.input
+package app.k9mail.feature.account.common.domain.input
 
 import app.k9mail.core.common.domain.usecase.validation.ValidationError
 import app.k9mail.core.common.domain.usecase.validation.ValidationResult
 
-class NumberInputField(
-    override val value: Long? = null,
+class BooleanInputField(
+    override val value: Boolean? = null,
     override val error: ValidationError? = null,
     override val isValid: Boolean = false,
-) : InputField<Long?> {
-
-    override fun updateValue(value: Long?): NumberInputField {
-        return NumberInputField(
+) : InputField<Boolean?> {
+    override fun updateValue(value: Boolean?): BooleanInputField {
+        return BooleanInputField(
             value = value,
             error = null,
             isValid = false,
         )
     }
 
-    override fun updateError(error: ValidationError?): NumberInputField {
-        return NumberInputField(
+    override fun updateError(error: ValidationError?): BooleanInputField {
+        return BooleanInputField(
             value = value,
             error = error,
             isValid = false,
         )
     }
 
-    override fun updateValidity(isValid: Boolean): NumberInputField {
+    override fun updateValidity(isValid: Boolean): BooleanInputField {
         if (isValid == this.isValid) return this
 
-        return NumberInputField(
+        return BooleanInputField(
             value = value,
             error = null,
             isValid = isValid,
         )
     }
 
-    override fun updateFromValidationResult(result: ValidationResult): NumberInputField {
+    override fun updateFromValidationResult(result: ValidationResult): BooleanInputField {
         return when (result) {
-            is ValidationResult.Success -> NumberInputField(
+            is ValidationResult.Success -> BooleanInputField(
                 value = value,
                 error = null,
                 isValid = true,
             )
 
-            is ValidationResult.Failure -> NumberInputField(
+            is ValidationResult.Failure -> BooleanInputField(
                 value = value,
                 error = result.error,
                 isValid = false,
@@ -55,7 +54,7 @@ class NumberInputField(
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
 
-        other as NumberInputField
+        other as BooleanInputField
 
         if (value != other.value) return false
         if (error != other.error) return false
@@ -70,6 +69,6 @@ class NumberInputField(
     }
 
     override fun toString(): String {
-        return "NumberInputField(value=$value, error=$error, isValid=$isValid)"
+        return "BooleanInputField(value=$value, error=$error, isValid=$isValid)"
     }
 }
