+++ feature/account/common/src/main/kotlin/app/k9mail/feature/account/common/domain/input/InputField.kt
@@ -1,4 +1,4 @@
-package app.k9mail.feature.account.setup.domain.input
+package app.k9mail.feature.account.common.domain.input
 
 import app.k9mail.core.common.domain.usecase.validation.ValidationError
 import app.k9mail.core.common.domain.usecase.validation.ValidationResult
