+++ feature/account/common/src/test/kotlin/app/k9mail/feature/account/common/domain/input/InputFieldTest.kt
@@ -1,4 +1,4 @@
-package app.k9mail.feature.account.setup.domain.input
+package app.k9mail.feature.account.common.domain.input
 
 import app.k9mail.core.common.domain.usecase.validation.ValidationError
 import app.k9mail.core.common.domain.usecase.validation.ValidationResult
