+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/ui/autodiscovery/item/ConfigurationApprovalItem.kt
@@ -3,8 +3,8 @@ package app.k9mail.feature.account.setup.ui.autodiscovery.item
 import androidx.compose.foundation.lazy.LazyItemScope
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import app.k9mail.feature.account.common.domain.input.BooleanInputField
 import app.k9mail.feature.account.common.ui.item.ListItem
-import app.k9mail.feature.account.setup.domain.input.BooleanInputField
 import app.k9mail.feature.account.setup.ui.autodiscovery.view.ConfigurationApprovalView
 
 @Composable
