+++ feature/account/setup/src/test/kotlin/app/k9mail/feature/account/setup/ui/options/AccountOptionsStateTest.kt
@@ -1,8 +1,8 @@
 package app.k9mail.feature.account.setup.ui.options
 
+import app.k9mail.feature.account.common.domain.input.StringInputField
 import app.k9mail.feature.account.setup.domain.entity.EmailCheckFrequency
 import app.k9mail.feature.account.setup.domain.entity.EmailDisplayCount
-import app.k9mail.feature.account.setup.domain.input.StringInputField
 import app.k9mail.feature.account.setup.ui.options.AccountOptionsContract.State
 import assertk.assertThat
 import assertk.assertions.isEqualTo
